repo: establish Tiny GPU Next architecture foundation
